requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
boto3==1.26.16
botocore==1.29.16
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
Faker==15.3.2
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
htbuilder==0.6.1
idna==3.4
iiif-prezi3==1.1.0
importlib-metadata==5.0.0
inflect==6.0.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
kiwisolver==1.4.4
langcodes==3.3.0
language-data==1.1
marisa-trie==0.7.8
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.0
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
st-annotated-text==3.0.0
streamlit==1.15.1
streamlit-aggrid==0.3.3
streamlit-camera-input-live==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.2
streamlit-faker==0.0.2
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
zipp==3.10.0
